requirements.txt
flask==2.3.3
sentence-transformers==2.2.2
chromadb==0.4.15
langchain==0.1.16
langchain-community==0.0.35
pypdf==3.17.1
python-dotenv==1.0.0
llama-cpp-python==0.2.48
youtube-transcript-api==0.6.1
beautifulsoup4==4.12.2
requests==2.31.0
Pillow==10.0.1
transformers==4.35.2
torch>=2.2.0
PyPDF2
huggingface_hub==0.16.4
waitress
psutil
python-dotenv 